[diff] Improve writing of diff
Only write a diff if the error is "Buffers differ"
Don't remove an old diff until there's a matching result.

+++ src/cli/commands/diff.ts
@@ -165,9 +165,9 @@ export const run = async (
             const outputDiffPath = site.out.absolute(
               `diff/${sanitizedName}/${sanitizedUrl}-${index}.diff.png`,
             );
-            if (result) {
+            if (result && result.errorMessage == "Buffers differ") {
               await fs.writeFile(outputDiffPath, result.diff);
-            } else if (await fileExists(outputDiffPath)) {
+            } else if (!result && (await fileExists(outputDiffPath))) {
               await fs.unlink(outputDiffPath);
             }
           }
